Changed Naming Conventions - 'tab component' renamed to Tabs - 'all tabs' renamed to DemoTabs - All instances of .js changed to .jsx - All file names converted to UpperCamelCase
