+++ diagnostic-ui/src/components/StartStopButton/StartStopButton.jsx
